--- app/myecommerce/src/app/login/index.tsx
@@ -1,7 +1,0 @@
-import LoginCard from "@/components/LoginCard"
-
-export default function Login() {
-  <div className="min-h-screen flex items-center justify-center bg-gray-900">
-    <LoginCard />
-  </div>
-}
